Directly add FIRE config files to the template
This makes the template simpler to understand compared to the
case of generating the configs at runtime.

+++ template.fire/run.sh
@@ -7,23 +7,20 @@ fi
 
 MATH=${{MATH:-math}}
 
+if ! which $MATH >/dev/null; then
+    echo "MATH must be set to the Mathematica CLI binary name" 1>&2
+    exit 1
+fi
+
 rm -f *.start *.lbases *.sbases
 rm -rf litered
 
-$MATH -script step1.m
-$MATH -script step2.m
-$MATH -script step3.m
+$MATH -script step1.m || exit 1
+$MATH -script step2.m || exit 1
+$MATH -script step3.m || exit 1
 
-env THREADS=2 \
-    sh ./fire-with-hint.sh \
-        -hint-variables "d->521,m2->44203,s12->46639,s23->52289" \
-        -hint-prime "1" \
-        -norm-variables "d,m2,s12,s23" \
-        -start "" \
-        -problem "1 |1,4|$PWD/basisx.sbases" \
-        -lbases "$PWD/basisx.lbases" \
-        -integrals "$PWD/integrals" \
-        -hint-output "$PWD/basisx.hint.tables" \
-        -norm-output "$PWD/basisx.tables"
+mkdir fire.hintdir
+$FIREPATH/bin/FIRE6p -c fire.hint || exit 1
+$FIREPATH/bin/FIRE6 -c fire || exit 1
 
-$MATH -script step4.m
+$MATH -script step4.m || exit 1
